requirements.txt
streamlit==1.37.1
openai==1.60.2
google-generativeai
requests==2.32.3
beautifulsoup4==4.12.3
playwright
Pillow
